refactor: move agents used calculation inside of the component that does the calculation
